PLAT-24909: Add ViewManager to other components (#10)
* apply ViewManager to Pickers

* refactor components into individual files

* use noAnimation instead of inanimate/animate

* update tests and add docs to Picker and RangePicker

* add custom propType validator for stepped numbers

* correct noAnimation docs

* rename reverse to reverseTransition

* fix propType ordering in Viewport

* update docs for noAnimation

* standardize around moon-button-height for Picker height

* add default value doc

* move all transition direction logic into PickerCore

Since we've added direction logic for wrapping, it makes sense to
extend that to determine the direction for any change since value will
always be an integer representing the relative position in the set.
This allows us to remove the logic from RangePicker and just set a
value-unique key to ensure React knows to re-render the children when
it changes.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)

Reviewed-By: Blake Stephens (blake.stephens@lge.com)
Integrated-By: Aaron Tam (aaron.tam@lge.com)