defer ViewManager children reconciliation until after setState
If new props came in before a prior setState re-rendered, the stateful
members (e.g. this.keysToLeave) could get out of sync with the actual
children that were present. This was further compounded by the change
to prepend new children on merge (by ENYO-3558) which didn't update the
drop logic which assumed the new children were appended.

The resolution is to move all the reconciliation into the setState
callback so we're updating the stateful members in sync with
this.state.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)