prevent setting un-focusable nodes as last-focused for container
The logic within container.isNavigable is more permissive and allows
sub-containers to pass in order to include them in the array of
spottable descendants.

In the case of the mouseover handler for spotlight, containers should
be omitted and only navigable elements included. This change prevents
settings containers as the last-focused for the purposes of mouseover.

+ optimization of getNavigableTarget to only consider the containers
  containing the target rather than all containers.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)