move spotPanel into handler to ensure a consistent ref between renders
The issue was introduced in 1.3.0 when `spotPanel` became a function returning a function `(autoFocus) => (node) => {}` instead of a function returning nothing `(node) => {}`. The side-effect of that change (coupled with Panel being an SFC) is that the function passed to `ref` on `<article>` is different every render which causes React to invoke it every render rather than only when the component is mounted.

This change combines the logic in the computed `spotOnRender` function and the `spotPanel` callback into a handler which will be the same ref across renders.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)