add ability to restore focus when spottable becomes enabled
If a container was focused explicitly (e.g. a Panel on render) but
did not contain any spottable children, spotlight is "lost" if the app
doesn't explicitly focus something.

This change remembers the attempted focus as the last active container
and adds the ability for Spottable to attempt to focus something within
that container (likely the Spottable itself) to re-establish spotlight.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)