ENYO-4685: Fix getLastContainer usage in spotlight
With the change to getLastContainer to return a falsey value for
inactive containers, we were not correctly handling its usage elsewhere
in spotlight.

This fixes the usage of that method to either use the "safe"
Spotlight.getActiveContainer() or new restoreFocus() which consolidates
the logic for finding the best focus target when none currently exists.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)