check for node rather than index range to restore focus in list
If the last focused index is within the rendered set but beyond the
visible set, the visible indices would not be updated to include it due
to scrolling the viewport to make it available. As a result, focus
would not be restored.

Replace the index checking with a querySelector call to try to find the
node and continue the restore focus logic if the node is found.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)