add imperative API to unmount all floating layers
When FloatingLayerDecorator is used within a subtree of the app, the
app/component may want to unmount all floating layers it owns (e.g. if
it is hiding all layers as in moonstone/VideoPlayer).

This change adds a new imperative API available, unmountFloatingLayers,
via context to support this.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)