Currently, when expandableList is opened, the first item is focused as default focus. UX requirement for 19Y TV is that selected item is focused when list is opened.
Basic behavior is that selected item is focused when list is opened.
If there's no selected, the first item is focused.
In the multi selection case, prop type of selected is number array. In this case, the first index of the array (index 0) is focused.
Once the index to be focused is determined, spotlight moves to that item based on data-index.

PLAT-48641 Change default focus in lists (First -> Selected)

Enact-DCO-1.0-Signed-off-by: Sangwook Lee (sangwook1203.lee@lge.com)