refactor binding support for handle
Previously, handle really only supporting binding itself to an object
but not the handler it would return. In order to support this use case,
which itself is useful for defining handlers at construction time and
binding them to an instance, some refactoring was required to cache and
evaluate different "this"-es to establish the caller.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)