Removed title->placeholder
If we really want to support this feature we can add it back later. In the mean time, it's safer to not include it, since it's a little presumptuous and users have direct access to add their own placeholder already. It's harder to take it away in the future than it is to add it.

Enact-DCO-1.0-Signed-off-by: Blake Stephens <blake.stephens@lge.com>