PLAT-52473: Add voice label in IncrementSlider
In current status, aria-label is declared in IconButton of IncrementSlider

But the label is included additional guidance text.

For example, incrementAriaLabel is "press ok button to increase the value"

But for voice control behavior, the button has to set "increment" label

In additionally, the label has to customized for app(ex. photovideo zoom slider)

Enact-DCO-1.0-Signed-off-by: Changgi Lee <changgi.lee@lge.com>