WRP-6257:Cell: add `grow` boolean prop to expand the size to its container
A `grow`able cell will expand to its maximum size according to the
remaining space of the container. This is used when you want to grow
the size of this cell so that it fills the container.

When combined with `shrink, `shrink` prop takes precedence over `grow`
prop and `grow` prop is simply ignored. When used in conjunction with
`size`, the size will be the minimum size, growing as necessary, to fit
the container.

A `grow`able Cell has a declaration of `flex: 1 1 100%` in CSS by
default. However, the `flex-basis` property is overridden by the
`var(--cell-size)` value passed from JS. And by combining `max-width:
none` and `flex-grow: 1` declaration, its size can be increased by the
size of its parent container.

Also, by the declaration of `flex: 1 1 100%`, the `flex-shrink`
property has a value of `1`, which means that its size can be reduced
as much as necessary, but because of `min-width: var(--cell-size)` and
`min-height: var(--cell-size)` declaration, it does not decrease
below the specified cell size.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>