WRP-6257:Cell: add `grow` boolean prop to expand size to its container
A `grow`able cell will expand to its maximum size, according to the
remaining space of the container. This is used when you want to grow
the size of this Cell so that it fills the container.

A `grow`able Cell has a declaration of `flex: 1 1 100%` in CSS by
default. However, the `flex-basis` property is overridden by the
`var(--cell-size)` value passed from JS. And by combining `max-width:
none` and `flex-grow: 1` declaration, its size can be increased by the
size of its parent container.

Also, by the declaration of `flex: 1 1 100%`, the `flex-shrink`
property have a value of `1`, which means that its size can be reduced
as much as necessary, but because of `min-width: var(--cell-size)` and
`min-height: var(--cell-size)` declaration, it does not descrease
below the specified cell size.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>