WRP-6257: Cell: Add `grow` prop to expand the size to its container (#3128)
* WRP-6257:Cell: add `grow` boolean prop to expand the size to its container

A `grow`able cell will expand to its maximum size according to the
remaining space of the container. This is used when you want to grow
the size of this cell so that it fills the container.

When combined with `shrink, `shrink` prop takes precedence over `grow`
prop and `grow` prop is simply ignored. When used in conjunction with
`size`, the size will be the minimum size, growing as necessary, to fit
the container.

A `grow`able Cell has a declaration of `flex: 1 1 100%` in CSS by
default. However, the `flex-basis` property is overridden by the
`var(--cell-size)` value passed from JS. And by combining `max-width:
none` and `flex-grow: 1` declaration, its size can be increased by the
size of its parent container.

Also, by the declaration of `flex: 1 1 100%`, the `flex-shrink`
property has a value of `1`, which means that its size can be reduced
as much as necessary, but because of `min-width: var(--cell-size)` and
`min-height: var(--cell-size)` declaration, it does not decrease
below the specified cell size.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>

* fixup! WRP-6257:Cell: add `grow` boolean prop to expand the size to its container

* WRP-6257:Layout: add unit tests for `grow` prop of `Cell`

Unit Test Cases
- `grow` CSS class is applied when no props (`shrink`, `size`) are
   given
- `grow` CSS class and `flex-basis` CSS style are applied when `grow`
   and `size` props are given
- `grow` CSS class is not applied when used with `shrink` prop

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>

* WRP-6257:sampler: add `grow` controllable Cell in Layout story

The thrid Cell is modified to change the `grow` and `size` props.
The part where `grow` is automatically activated is still checkable
when `shrink` prop is disabled in the second Cell.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>

* fixup! WRP-6257:sampler: add `grow` controllable Cell in Layout story

* fixup! WRP-6257:sampler: add `grow` controllable Cell in Layout story

* fixup! fixup! WRP-6257:Cell: add `grow` boolean prop to expand the size to its container

* fixup! WRP-6257:Cell: add `grow` boolean prop to expand the size to its container