WRP-11806:platform: update regular expression for Edge browser UA
According to the documents
- https://learn.microsoft.com/en-us/microsoft-edge/web-platform/user-agent-guidance
- https://www.whatismybrowser.com/guides/the-latest-user-agent/edge.

The identifier token of Edge browsers can be one of the following
user agents.

|          Platform |     UA |
|-------------------|--------|
|           Windows |    Edg |
|             macOS |    Edg |
|           Android |   EdgA |
|               iOS | EdgiOS |
| Windows 10 Mobile |   Edge |
|          Xbox One |   Edge |

The identifier token can be `Edg`, `Edge`, `EdgA`, `EdgiOS`.

In the case of `EdgA` and `Edge`, `Android` is also included in the UA
string, so `edge` should be checked before `androidChrome`.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>